+++ tools/file.h
@@ -6,11 +6,11 @@ extern "C" {
 #endif
 
 extern int file_read_content(const char *filename,char *buffer,
-									int size, long offset);
+									size_t size, long offset);
 extern int file_write_content(const char *filename,char *buffer,
-									int size, long offset);
+									size_t size, long offset);
 extern int file_append_content(const char *filename,char *buffer,
-									int size);
+									size_t size);
 extern int file_get_size(const char *filename);
 extern char *file_get_striped_name(const char *filename);
